+++ cdborm/tests/__init__.py
@@ -14,6 +14,7 @@ from .multidatabase import MultiDatabaseTest
 from .multidatabase_o2o import MultiDatabaseO2OTest
 from .multidatabase_o2m import MultiDatabaseO2MTest
 from .multidatabase_m2m import MultiDatabaseM2MTest
+from .no_init_db import NoInitDbTest
 
 all_test_cases = [
     ModelTest,
@@ -30,6 +31,7 @@ all_test_cases = [
     MultiDatabaseO2MTest,
     MultiDatabaseM2MTest,
     ListFieldTest,
+    NoInitDbTest,
 ]