+++ History.md
@@ -5,6 +5,7 @@
   * Fixed element name bug
   * Fixed problem with infinite loop made by relations.
   * Init now expects objects for relation, not id's.
+  * Fixed choosing wrong database in Model.all
 
 0.1.0 / 2012-12-15
 ==================